Flytter kode til service for å få transactional på opprettelse av tasl
